He actualizado el ejercicio 5

+++ ejercicio5.js
@@ -1,5 +1,5 @@
-const peso = parseInt(prompt("¿Cuál es tu peso?"));
-const altura = parseInt(prompt("¿Cuál es tu altura?"));
+const peso = parseFloat(prompt("¿Cuál es tu peso?"));
+const altura = parseFloat(prompt("¿Cuál es tu altura?"));
 
 const imc = peso/altura**2;
